Move ProfileValidationException to exception package
