+++ src/main/java/org/duraspace/bagit/exception/ProfileValidationException.java
@@ -2,7 +2,7 @@
  * The contents of this file are subject to the license and copyright detailed
  * in the LICENSE and NOTICE files at the root of the source tree.
  */
-package org.duraspace.bagit;
+package org.duraspace.bagit.exception;
 
 /**
  * A profile validation exception.
